// VIS_Application/app/Model/Masters/Configuration/MacIdConfiguration.ts
export interface IMacIdConfiguration
{
    Id: number;
    EmployeeId: number;
    Employee_Name: string;

    MacID1: string;
    MacID2: string;
    MacID3: string;
    MacID4: string;
    MacID5: string;
    MacID6: string;
    MacID: string;

    IP1: string;
    IP2: string;
    IP3: string;
    IP4: string;
    IPAddress: string;

    RequestedDate: Date;
    ApprovalDate: Date;
    ApprovedBy: string; 

    Version1: string;
    Version2: string;
    Version3: string;
    Version: string;

    Active: boolean;
    OfficeMacId: boolean; 
    Description: string;
    CreatedOn: Date;
    CreatedBy: string;
    UpdatedOn: Date;
    UpdatedBy: string
    IsActive: boolean;

}